+++ src/ethereum/forks/amsterdam/fork_types.py
@@ -13,34 +13,16 @@ Types reused throughout the specification, which are specific to Ethereum.
 
 from dataclasses import dataclass
 
-from ethereum_rlp import rlp
-from ethereum_types.bytes import Bytes, Bytes256
+from ethereum_types.bytes import Bytes256
 from ethereum_types.frozen import slotted_freezable
 from ethereum_types.numeric import U8, U64, U256
 
-from ethereum.crypto.hash import Hash32, keccak256
-from ethereum.state import Account, Address
+from ethereum.crypto.hash import Hash32
+from ethereum.state import Address
 
 VersionedHash = Hash32
 
 Bloom = Bytes256
-
-
-def encode_account(raw_account_data: Account, storage_root: Bytes) -> Bytes:
-    """
-    Encode `Account` dataclass.
-
-    Storage is not stored in the `Account` dataclass, so `Accounts` cannot be
-    encoded without providing a storage root.
-    """
-    return rlp.encode(
-        (
-            raw_account_data.nonce,
-            raw_account_data.balance,
-            storage_root,
-            keccak256(raw_account_data.code),
-        )
-    )
 
 
 @slotted_freezable
